Add FindLSB CMake module for Linux distro info.
Different distros may require different packaging (for example
if we check for installed rpms or debs), so we want to detect
the distribution we're building on. May also want to encode
this into platform checks and the package name.

Currently runs, though we don't use the output yet. That
needs testing on several platforms to get consistency.